Rotate images appropriately for posters too
